+++ src/MapWindowTraffic.cpp
@@ -45,7 +45,7 @@ Copyright_License {
 #include "Screen/Layout.hpp"
 #include "StringUtil.hpp"
 
-#define fSnailColour(cv) max(0,min((short)(NUMSNAILCOLORS-1), (short)((cv+1.0)/2.0*NUMSNAILCOLORS)))
+#define fSnailColour(cv) max((short)0, min((short)(NUMSNAILCOLORS-1), (short)((cv+1.0)/2.0*NUMSNAILCOLORS)))
 
 /**
  * Draws the FLARM traffic icons onto the given canvas
